Move otel status serialization to a separate package
Backport of the serialization refactoring from b3a6511126 (PR #11856).
Moves SerializableStatus, SerializableEvent, and related types and
functions from the manager package to internal/pkg/otel/status, making
them reusable by other packages.
