+++ internal/pkg/otel/manager/execution_subprocess.go
@@ -18,9 +18,11 @@ import (
 	"go.opentelemetry.io/collector/component"
 	"gopkg.in/yaml.v3"
 
+	"github.com/elastic/elastic-agent/internal/pkg/otel/status"
+
 	componentmonitoring "github.com/elastic/elastic-agent/internal/pkg/agent/application/monitoring/component"
 
-	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/status"
+	otelstatus "github.com/open-telemetry/opentelemetry-collector-contrib/pkg/status"
 	"go.opentelemetry.io/collector/component/componentstatus"
 	"go.opentelemetry.io/collector/confmap"
 	"go.uber.org/zap/zapcore"
@@ -76,7 +78,7 @@ type subprocessExecution struct {
 
 // startCollector starts a supervised collector and monitors its health. Process exit errors are sent to the
 // processErrCh channel. Other run errors, such as not able to connect to the health endpoint, are sent to the runErrCh channel.
-func (r *subprocessExecution) startCollector(ctx context.Context, baseLogger *logger.Logger, logger *logger.Logger, cfg *confmap.Conf, processErrCh chan error, statusCh chan *status.AggregateStatus, forceFetchStatusCh chan struct{}) (collectorHandle, error) {
+func (r *subprocessExecution) startCollector(ctx context.Context, baseLogger *logger.Logger, logger *logger.Logger, cfg *confmap.Conf, processErrCh chan error, statusCh chan *otelstatus.AggregateStatus, forceFetchStatusCh chan struct{}) (collectorHandle, error) {
 	if cfg == nil {
 		// configuration is required
 		return nil, errors.New("no configuration provided")
@@ -150,7 +152,7 @@ func (r *subprocessExecution) startCollector(ctx context.Context, baseLogger *lo
 		defer func() {
 			close(healthCheckDone)
 		}()
-		currentStatus := aggregateStatus(componentstatus.StatusStarting, nil)
+		currentStatus := status.AggregateStatus(componentstatus.StatusStarting, nil)
 		r.reportSubprocessCollectorStatus(ctx, statusCh, currentStatus)
 
 		// specify a max duration of not being able to get the status from the collector
@@ -178,7 +180,7 @@ func (r *subprocessExecution) startCollector(ctx context.Context, baseLogger *lo
 			} else {
 				maxFailuresTimer.Reset(maxFailuresDuration)
 				removeManagedHealthCheckExtensionStatus(statuses, r.healthCheckExtensionID)
-				if !compareStatuses(currentStatus, statuses) {
+				if !status.CompareStatuses(currentStatus, statuses) {
 					currentStatus = statuses
 					r.reportSubprocessCollectorStatus(procCtx, statusCh, statuses)
 				}
@@ -194,11 +196,11 @@ func (r *subprocessExecution) startCollector(ctx context.Context, baseLogger *lo
 			case <-healthCheckPollTimer.C:
 				healthCheckPollTimer.Reset(healthCheckPollDuration)
 			case <-maxFailuresTimer.C:
-				failedToConnectStatuses := aggregateStatus(
+				failedToConnectStatuses := status.AggregateStatus(
 					componentstatus.StatusRecoverableError,
 					errors.New("failed to connect to collector"),
 				)
-				if !compareStatuses(currentStatus, failedToConnectStatuses) {
+				if !status.CompareStatuses(currentStatus, failedToConnectStatuses) {
 					currentStatus = statuses
 					r.reportSubprocessCollectorStatus(procCtx, statusCh, statuses)
 				}
@@ -245,17 +247,17 @@ func (r *subprocessExecution) startCollector(ctx context.Context, baseLogger *lo
 }
 
 // cloneCollectorStatus creates a deep copy of the provided AggregateStatus.
-func cloneCollectorStatus(aStatus *status.AggregateStatus) *status.AggregateStatus {
+func cloneCollectorStatus(aStatus *otelstatus.AggregateStatus) *otelstatus.AggregateStatus {
 	if aStatus == nil {
 		return nil
 	}
 
-	st := &status.AggregateStatus{
+	st := &otelstatus.AggregateStatus{
 		Event: aStatus.Event,
 	}
 
 	if len(aStatus.ComponentStatusMap) > 0 {
-		st.ComponentStatusMap = make(map[string]*status.AggregateStatus, len(aStatus.ComponentStatusMap))
+		st.ComponentStatusMap = make(map[string]*otelstatus.AggregateStatus, len(aStatus.ComponentStatusMap))
 		for k, cs := range aStatus.ComponentStatusMap {
 			st.ComponentStatusMap[k] = cloneCollectorStatus(cs)
 		}
@@ -264,7 +266,7 @@ func cloneCollectorStatus(aStatus *status.AggregateStatus) *status.AggregateStat
 	return st
 }
 
-func (r *subprocessExecution) reportSubprocessCollectorStatus(ctx context.Context, statusCh chan *status.AggregateStatus, collectorStatus *status.AggregateStatus) {
+func (r *subprocessExecution) reportSubprocessCollectorStatus(ctx context.Context, statusCh chan *otelstatus.AggregateStatus, collectorStatus *otelstatus.AggregateStatus) {
 	// we need to clone the status to prevent any mutation on the receiver side
 	// affecting the original ref
 	clonedStatus := cloneCollectorStatus(collectorStatus)
@@ -302,7 +304,7 @@ func (r *subprocessExecution) getCollectorPorts() (healthCheckPort int, metricsP
 	return healthCheckPort, metricsPort, nil
 }
 
-func removeManagedHealthCheckExtensionStatus(status *status.AggregateStatus, healthCheckExtensionID string) {
+func removeManagedHealthCheckExtensionStatus(status *otelstatus.AggregateStatus, healthCheckExtensionID string) {
 	extensions, exists := status.ComponentStatusMap["extensions"]
 	if !exists {
 		return
